Added epigraph
Added a new epigraph command `iowaEpigraph` this has to do with the fact
that memoir already defines an epigraph command. Borrowed their idea
to make it a command which typesets everything (including the source)
instead of our first idea of an enviroment.

Currently if the quote is too short this looks funny. It's also not so
clear if what it does counts as being centered.