Defined a new chapter header style.
Also changed the behavior of the default section heading, just making it
more normal sized. As far as I can tell, there doesn't seem to be a way
to give it its own defined style, so it can't "easily" be changed aside
from editing the .cls file. It can be omitted, but the default section
header is just huge.