Initial release target added
Added a make target which creates a release. The release consists of the
example pdf and a zip file of what was used to created it. The idea is
most people will just download this from the github page and modify the
example to get their thesis off the ground.