Handle options by default
To reduce the amount of noise in the example document the class file
now checks if the use specified how many sides to have or what side to
open on. I don't really want to explain those options and I'm trying to
minimize options which the end user doesn't understand. 11pt and
letterpaper felt obvious enough so I left em in.